backend/requirements/prod.txt
# Core
pytz==2021.1
Django==3.2.3
gunicorn==20.1.0
newrelic==6.4.0.157
django-dotenv==1.4.2

# For the persistence stores
psycopg2==2.8.6
redis==3.5.3

# Model Tools
django-model-utils==4.1.1
django_unique_upload==0.2.1

# Rest apis
djangorestframework==3.12.4
djangorestframework-camel-case=1.2.0
djangorestframework-simplejwt==4.7.1
Markdown==3.3.4
drf-yasg==1.20.0
django-filter==2.4.0
django-cors-headers==3.7.0
django-rest-swagger==2.2.0
